Move session init code into server-config.
