+++ lib/create-app.js
@@ -1,9 +1,7 @@
 module.exports = createApp
 
 const express = require('express')
-const session = require('express-session')
 const handlebars = require('express-handlebars')
-const uuid = require('uuid')
 const { addLdpMiddleware } = require('./data-storage/api')
 const SolidHost = require('./solid-host')
 const { AccountManager } = require('./account-mgmt/account-manager')
@@ -15,7 +13,6 @@ const ShareRequest = require('./authorization/share-request')
 const errorPages = require('./error-pages')
 const config = require('./server-config')
 const defaults = require('./defaults')
-const { logger } = require('./logger')
 const path = require('path')
 const bodyParser = require('body-parser')
 const { ldpRequestHandler } = require('./data-storage/api')
@@ -122,30 +119,13 @@ function initViews (app, configPath) {
  * @param storage {StorageManager}
  */
 function initWebId (argv, app, storage) {
+  const { host } = argv
   config.ensureWelcomePage({ argv, storage })
 
   // Store the user's session key in a cookie
   // (for same-domain browsing by people only)
   const useSecureCookies = !!argv.sslKey // use secure cookies when over HTTPS
-  const sessionHandler = session(sessionSettings(useSecureCookies, argv.host))
-  app.use((req, res, next) => {
-    sessionHandler(req, res, () => {
-      // Reject cookies from third-party applications.
-      // Otherwise, when a user is logged in to their Solid server,
-      // any third-party application could perform authenticated requests
-      // without permission by including the credentials set by the Solid server.
-      const origin = req.headers.origin
-      const userId = req.session.userId
-      if (!argv.host.allowsSessionFor(userId, origin)) {
-        logger.warn(`Rejecting session for ${userId} from ${origin}`)
-        // Destroy session data
-        delete req.session.userId
-        // Ensure this modified session is not saved
-        req.session.save = (done) => done()
-      }
-      next()
-    })
-  })
+  config.initSessionHandler({ app, useSecureCookies, host })
 
   const accountManager = AccountManager.from({
     authMethod: argv.auth,
@@ -186,35 +166,6 @@ function initAuthentication (app, argv) {
     throw new Error(`Unsupported authentication scheme: ${auth}`)
   }
   authenticationApi[auth].initialize(app, argv)
-}
-
-/**
- * Returns a settings object for Express.js sessions.
- *
- * @param secureCookies {boolean}
- * @param host {SolidHost}
- *
- * @return {Object} `express-session` settings object
- */
-function sessionSettings (secureCookies, host) {
-  const sessionSettings = {
-    secret: uuid.v1(),
-    saveUninitialized: false,
-    resave: false,
-    rolling: true,
-    cookie: {
-      maxAge: 24 * 60 * 60 * 1000
-    }
-  }
-  // Cookies should set to be secure if https is on
-  if (secureCookies) {
-    sessionSettings.cookie.secure = true
-  }
-
-  // Determine the cookie domain
-  sessionSettings.cookie.domain = host.cookieDomain
-
-  return sessionSettings
 }
 
 /**
