+++ lib/server-config.js
@@ -6,6 +6,8 @@
 
 const fs = require('fs-extra')
 const path = require('path')
+const uuid = require('uuid')
+const session = require('express-session')
 const { AccountTemplate, processHandlebarFile } = require('./account-mgmt/account-template')
 const LegacyResourceMapper = require('./data-storage/ldp-backend-fs/legacy-resource-mapper')
 const { StorageManager } = require('./storage-manager')
@@ -192,13 +194,68 @@ function initTemplateDirs (configPath) {
   }
 }
 
+function initSessionHandler ({ app, useSecureCookies, host }) {
+  // Store the user's session key in a cookie
+  // (for same-domain browsing by people only)
+  const sessionHandler = session(sessionSettings(useSecureCookies, host))
+  app.use((req, res, next) => {
+    sessionHandler(req, res, () => {
+      // Reject cookies from third-party applications.
+      // Otherwise, when a user is logged in to their Solid server,
+      // any third-party application could perform authenticated requests
+      // without permission by including the credentials set by the Solid server.
+      const origin = req.headers.origin
+      const userId = req.session.userId
+      if (!host.allowsSessionFor(userId, origin)) {
+        logger.warn(`Rejecting session for ${userId} from ${origin}`)
+        // Destroy session data
+        delete req.session.userId
+        // Ensure this modified session is not saved
+        req.session.save = (done) => done()
+      }
+      next()
+    })
+  })
+}
+
+/**
+ * Returns a settings object for Express.js sessions.
+ *
+ * @param secureCookies {boolean}
+ * @param host {SolidHost}
+ *
+ * @return {Object} `express-session` settings object
+ */
+function sessionSettings (secureCookies, host) {
+  const sessionSettings = {
+    secret: uuid.v1(),
+    saveUninitialized: false,
+    resave: false,
+    rolling: true,
+    cookie: {
+      maxAge: 24 * 60 * 60 * 1000
+    }
+  }
+  // Cookies should set to be secure if https is on
+  if (secureCookies) {
+    sessionSettings.cookie.secure = true
+  }
+
+  // Determine the cookie domain
+  sessionSettings.cookie.domain = host.cookieDomain
+
+  return sessionSettings
+}
+
 module.exports = {
   ensureDirCopyExists,
   ensureWelcomePage,
   hostConfigFor,
   initConfigDir,
   initDefaultViews,
+  initSessionHandler,
   initStorage,
   initTemplateDirs,
-  printDebugInfo
+  printDebugInfo,
+  sessionSettings
 }
